requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
deprecation==2.1.0
dnspython==2.7.0
email_validator==2.2.0
executing==2.0.1
fastapi==0.115.2
fastapi-cli==0.0.5
frozenlist==1.4.1
gotrue==2.9.2
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httptools==0.6.2
httpx==0.27.2
hyperframe==6.0.1
idna==3.7
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kombu==5.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
nest-asyncio==1.6.0
outcome==1.3.0.post0
packaging==24.1
parso==0.8.4
platformdirs==4.2.2
pluggy==1.5.0
postgrest==0.17.1
prompt_toolkit==3.0.47
propcache==0.2.0
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PySocks==1.7.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.1.0
realtime==2.0.5
redis==5.2.0
requests==2.32.3
rich==13.9.2
rq==2.0.0
selenium==4.23.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.40.0
storage3==0.8.1
StrEnum==0.4.15
supabase==2.9.0
supafunc==0.6.1
tornado==6.4.1
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
uvicorn==0.32.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
wsproto==1.2.0
yarl==1.15.4
